Finished authentication API

+++ backend/controllers/goalController.js
@@ -1,6 +1,8 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 
 /**
  * @description Get goals
@@ -8,7 +10,7 @@ const Goal = require('../models/goalModel')
  * @access Private
  */
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -26,6 +28,7 @@ const getGoals = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -42,6 +45,20 @@ const getGoals = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Check if logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(403)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -63,6 +80,21 @@ const getGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+    
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Check if logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(403)
+        throw new Error('User not authorized')
+    }
+
 
     await Goal.findByIdAndDelete(goal)
